- fixed a typo where /xdid -d wouldnt work for comboex - added a check for valid N input for /xdid -d
git-svn-id: http://dcx.googlecode.com/svn@59 17a06da3-8c20-fa79-550f-e9588e724b77

+++ Classes/dcxcomboex.cpp
@@ -343,12 +343,14 @@ void DcxComboEx::parseCommandRequest( TString & input ) {
 	 }
   }
   // xdid -d [NAME] [ID] [SWITCH] [N]
-  else if ( flags.switch_flags[2] && numtok > 3 ) {
-
+  else if ( flags.switch_flags[3] && numtok > 3 ) {
     int nItem = atoi( input.gettok( 4, " " ).to_chr( ) ) - 1;
 
-    if ( nItem > -1 )
+	 if (nItem > -1 && nItem < this->getCount())
       this->deleteItem( nItem );
+
+	 if (!this->getCount())
+		 this->redrawWindow();
   }
   // xdid -r [NAME] [ID] [SWITCH]
   else if ( flags.switch_flags[17] ) {
